ddl sql for different storage.

+++ packages/watchmen-storage-oracle/meta-scripts/16.5.19/00001-change_data_record.ddl.sql
@@ -1,2 +1,2 @@
 ALTER TABLE change_data_record
-    ADD status BIGINT;
+    ADD status NUMBER(1);
