+++ packages/watchmen-storage-oracle/meta-scripts/16.5.19/00006-scheduled_task_history.ddl.sql
@@ -1,2 +1,2 @@
 ALTER TABLE scheduled_task_history
-    ADD status BIGINT;
+    ADD status NUMBER(1);
